chore: refactored output plugin to follow other plugin structures
